Moving operations list to its own page. Fixes #61
