+++ routes.go
@@ -4,6 +4,7 @@ import "net/http"
 
 func init() {
 	http.HandleFunc("/app", App)
+	http.HandleFunc("/app/ops", AppOps)
 	http.HandleFunc("/inv", Inv)
 	http.HandleFunc("/op", Op)
 }
